cvailor-backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "seeds"]

[project]
name = "cvailor-backend"
version = "0.1.0"
description = "Cvailor AI-powered CV builder backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",

    # Database
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # Validation & settings
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
    "python-multipart>=0.0.9",

    # Redis & background tasks
    "celery[redis]>=5.4.0",
    "redis>=5.0.4",
    "flower>=2.0.1",

    # AI
    "anthropic>=0.25.0",
    "openai>=1.30.0",
    "tiktoken>=0.7.0",

    # File handling / export
    "boto3>=1.34.0",
    "aiofiles>=23.2.1",

    # HTTP client
    "httpx>=0.27.0",

    # Utilities
    "python-slugify>=8.0.4",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
